automatically backed up by learn

+++ index.js
@@ -47,7 +47,7 @@ function makeInt(n)
 
 function preserveDecimal(n)
 {
-  parseInt(n)
+  parseFloat(n)
 }
